requirements.txt
python-dotenv>=1.0.1,<2.0.0
langchain>=0.2.0,<0.3.0
langchain-openai>=0.1.0,<0.2.0
langgraph>=0.2.0,<0.3.0
grandalf>=0.8,<0.9.0
sendgrid>=6.11.0,<7.0.0
streamlit>=1.38.0,<2.0.0
serpapi>=0.1.5,<0.2.0